Edit Order modal update
Another task, exactly the same as we've done with Add Purchase modal.
In the Edit Order modal when I enter the item in the line from the "Item Name" or "Mijn Tafel Name" drop-down menu I want the data in the Price (€) and Category columns to be automatically populated. The data should be taken from the most recent purchase transaction.